requirements.txt
chromadb==0.5.5
tqdm==4.66.4
python-dotenv==1.0.1
PyPDF2==3.0.1
requests==2.32.3
nltk==3.8.1
gradio==4.40.0
